fix: 네이버 인증 오류 해소
