+++ src/main/java/com/example/WonkaoTalk/domain/auth/dto/OAuth/GoogleUserInfo.java
@@ -1,25 +1,21 @@
-package com.example.WonkaoTalk.domain.auth.dto;
+package com.example.WonkaoTalk.domain.auth.dto.OAuth;
 
 import java.util.Map;
 import lombok.AllArgsConstructor;
 
 @AllArgsConstructor
-public class NaverUserInfo implements OAuth2UserInfo {
+public class GoogleUserInfo implements OAuth2UserInfo {
 
   private Map<String, Object> attributes;
 
-//  public NaverUserInfo(Map<String, Object> attributes) {
-//    this.attributes = (Map<String, Object>) attributes.get("response");
-//  }
-
   @Override
   public String getProvider() {
-    return "naver";
+    return "google";
   }
 
   @Override
   public String getProviderId() {
-    return (String) attributes.get("id");
+    return (String) attributes.get("sub");
   }
 
   @Override
@@ -31,4 +27,5 @@ public class NaverUserInfo implements OAuth2UserInfo {
   public String getName() {
     return (String) attributes.get("name");
   }
+
 }
