+++ src/main/java/com/example/WonkaoTalk/domain/auth/dto/OAuth/NaverUserInfo.java
@@ -1,21 +1,23 @@
-package com.example.WonkaoTalk.domain.auth.dto;
+package com.example.WonkaoTalk.domain.auth.dto.OAuth;
 
 import java.util.Map;
-import lombok.AllArgsConstructor;
 
-@AllArgsConstructor
-public class GoogleUserInfo implements OAuth2UserInfo {
+public class NaverUserInfo implements OAuth2UserInfo {
 
   private Map<String, Object> attributes;
 
+  public NaverUserInfo(Map<String, Object> attributes) {
+    this.attributes = (Map<String, Object>) attributes.get("response");
+  }
+
   @Override
   public String getProvider() {
-    return "google";
+    return "naver";
   }
 
   @Override
   public String getProviderId() {
-    return (String) attributes.get("sub");
+    return (String) attributes.get("id");
   }
 
   @Override
@@ -27,5 +29,4 @@ public class GoogleUserInfo implements OAuth2UserInfo {
   public String getName() {
     return (String) attributes.get("name");
   }
-
 }
